refactor: remote package in data module
