+++ trends_data/src/commonMain/kotlin/net/thechance/mena/trends/data/remote/dto/RemotePaginationResponse.kt
@@ -1,4 +1,4 @@
-package net.thechance.mena.trends.data.dto
+package net.thechance.mena.trends.data.remote.dto
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
