+++ trends_data/src/commonMain/kotlin/net/thechance/mena/trends/data/remote/repository/ReelsRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package net.thechance.mena.trends.data.repository
+package net.thechance.mena.trends.data.remote.repository
 
 import io.ktor.client.HttpClient
 import io.ktor.client.request.delete
@@ -21,21 +21,21 @@ import kotlinx.coroutines.flow.SharedFlow
 import kotlinx.io.buffered
 import net.thechance.mena.trends.data.di.TrendDataModule.Companion.DEFAULT_CLIENT_NAME
 import net.thechance.mena.trends.data.di.TrendDataModule.Companion.UPLOAD_CLIENT_NAME
-import net.thechance.mena.trends.data.dto.ReelDto
-import net.thechance.mena.trends.data.dto.RemotePaginationResponse
-import net.thechance.mena.trends.data.dto.ReelPathUrlsDto
-import net.thechance.mena.trends.data.dto.UpdateReelRequestDTO
-import net.thechance.mena.trends.data.dto.UploadReelResponse
-import net.thechance.mena.trends.data.mapper.toEntity
-import net.thechance.mena.trends.data.mapper.toReelUrls
+import net.thechance.mena.trends.data.remote.dto.ReelDto
+import net.thechance.mena.trends.data.remote.dto.ReelPathUrlsDto
+import net.thechance.mena.trends.data.remote.dto.RemotePaginationResponse
+import net.thechance.mena.trends.data.remote.dto.UpdateReelRequestDTO
+import net.thechance.mena.trends.data.remote.dto.UploadReelResponse
+import net.thechance.mena.trends.data.remote.mapper.toEntity
+import net.thechance.mena.trends.data.remote.mapper.toReelUrls
 import net.thechance.mena.trends.data.util.NetworkEndpoint.LIKE_REEL_ENDPOINT
 import net.thechance.mena.trends.data.util.NetworkEndpoint.PAGE_PARAMETER
 import net.thechance.mena.trends.data.util.NetworkEndpoint.PROFILE_REELS_ENDPOINT
 import net.thechance.mena.trends.data.util.NetworkEndpoint.REELS_FEED_ENDPOINT
+import net.thechance.mena.trends.data.util.NetworkEndpoint.REFRESH_REEL_ENDPOINT
 import net.thechance.mena.trends.data.util.NetworkEndpoint.THUMBNAIL_ENDPOINT
 import net.thechance.mena.trends.data.util.NetworkEndpoint.TRENDS_PATH
 import net.thechance.mena.trends.data.util.NetworkEndpoint.VIEW_REEL_ENDPOINT
-import net.thechance.mena.trends.data.util.NetworkEndpoint.REFRESH_REEL_ENDPOINT
 import net.thechance.mena.trends.data.util.NetworkKeys.THUMBNAIL
 import net.thechance.mena.trends.data.util.NetworkKeys.THUMBNAIL_MIME_TYPE
 import net.thechance.mena.trends.data.util.NetworkKeys.VIDEO
