+++ trends_data/src/commonTest/kotlin/net/thechance/mena/trends/data/repository/ReelRepositoryImplTest.kt
@@ -9,6 +9,7 @@ import io.ktor.client.HttpClient
 import io.ktor.http.HttpStatusCode
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.test.runTest
+import net.thechance.mena.trends.data.remote.repository.ReelsRepositoryImpl
 import net.thechance.mena.trends.data.repository.util.VideoFileHandlerMock
 import net.thechance.mena.trends.data.repository.util.addViewReelResponse
 import net.thechance.mena.trends.data.repository.util.createReelsHttpClient
